requirements.txt
python-dotenv
sounddevice
numpy
requests
google-cloud-speech
google-generativeai
arabic-reshaper
python-bidi
python-vlc
scipy